add noti when use old cv
