+++ pages/groups/[id]/register-is-mentor.vue
@@ -1,8 +1,8 @@
 <template>
   <BContainer class="pt-3">
     <BRow>
-      <BCol class="pb-3">
-        <h5>>> Đăng ký là người hướng dẫn của nhóm</h5>
+      <BCol class="pb-4">
+        <h5><BIconChevronRight /><BIconChevronRight /> Đăng ký là người hướng dẫn của nhóm</h5>
       </BCol>
     </BRow>
     <BRow>
@@ -33,21 +33,25 @@
           <p class="mb-0"> {{ index + 1 }}. {{ member.full_name }} _ Khoa: {{ member.faculty }}</p>
         </div>
         <div class="mt-5 register">
-          <h5 class="text-center" for="">Thông tin kiểm duyệt đăng ký là người hướng dẫn:</h5>
-          <p class="notice-success" v-if="statusShow === 2">
-            <span>Đăng ký đã thực hiện thành công!.</span>
-            Nhà trường sẽ xem xét thông tin bạn đăng ký, theo dõi email để cập nhật thông tin nhé.</p>
-          <div v-if="statusShow === 1">
-            <label for="" class="pb-3 pt-2">Tôi đã tìm hiểu kỹ về Thông báo tuyển chọn sinh viên tham gia hướng dẫn, hỗ
+          <h5 class="text-center mb-3" for="">Thông tin kiểm duyệt đăng ký là người hướng dẫn:</h5>
+          <p class="notice-success mb-2" v-if="statusShow === 2">
+            <span>Bạn đã đăng ký làm người hướng dẫn của nhóm!.</span>
+            Nhà trường sẽ xem xét thông tin bạn đăng ký, theo dõi email để cập nhật thông tin nhé.
+            <span class="pt-4 mb-0">Thông tin của bạn:</span>
+          </p>
+          <div>
+            <label for="" class="mb-3">Tôi đã tìm hiểu kỹ về Thông báo tuyển chọn sinh viên tham gia hướng dẫn, hỗ
               trợ học tập (Mentor) của Trường. Tôi nhận thấy bản thân đáp ứng các tiêu chuẩn của Mentor và đăng ký làm
               Mentor cho học phần {{ group.subject }}</label>
-            <form @submit.prevent="submit">
+              <form @submit.prevent="submit">
               <div role="group">
-                <label>1. Thành tích bạn đạt được ở môn học này. <span>Bỏ hết vào một thư mục và chia sẻ link đường dẫn ở
-                    chế độ xem public.</span></label>
+                <label>1. Thành tích bạn đạt được ở môn học này.
+                  <span v-if="beforeShow === 0">Bỏ hết vào một thư mục và chia sẻ link đường dẫn ở chế độ xem public.</span>
+                  <span v-if="beforeShow === 1">Bạn đã lưu thông tin thành tích cho môn học {{group.subject}} ở lần đăng ký trước.</span>
+                </label>
                 <BFormInput v-model="register_inform.cv_link"
-                  :state="validationErrorMessages.cv_link === undefined ? null : false" placeholder="Link thành tích" trim
-                  required class="" />
+                  :state="validationErrorMessages.cv_link === undefined ? null : false" placeholder="Link thành tích"
+                  trim required class="" />
                 <BFormInvalidFeedback>
                   <ValidationErrorMessage :messages="validationErrorMessages.cv_link" />
                 </BFormInvalidFeedback>
@@ -64,7 +68,10 @@
               </div>
 
               <div role="group">
-                <label for="">3. Tài khoản ngân hàng?<span>Bao gồm số tài khoản và tên ngân hàng.</span></label>
+                <label for="">3. Tài khoản ngân hàng?
+                  <span v-if="beforeShow === 0">Bao gồm số tài khoản và tên ngân hàng.</span>
+                  <span v-if="beforeShow === 1">Bạn đã lưu thông tin tài khoản ở lần đăng ký trước.</span>
+                </label>
                 <BFormTextarea v-model="register_inform.smart_banking"
                   :state="validationErrorMessages.smart_banking === undefined ? null : false"
                   aria-describedby="input-live-help input-live-feedback" placeholder="Tài khoản ngân hàng" trim required
@@ -94,9 +101,15 @@
                 <span class="confirm-error" v-if="showConfirmError">Bạn phải đảm bảo thông tin trên!</span>
               </div>
 
-              <div class="text-end">
+              <div class="text-end" v-if="statusShow === 1">
                 <SubmitButton class="mt-3 submit-button" :isDisabled="isDisabledButton" :content="'Đăng ký tham gia'"
                   :color="'rgb(63 88 120)'" />
+              </div>
+              <div class="text-end" v-if="statusShow === 2">
+                <SubmitButton class="mt-3 me-3 submit-button" :isDisabled="isDisabledButton"
+                  :content="'Chỉnh sửa thông tin'" :color="'rgb(23 131 27)'" @click.prevent="update" />
+                <SubmitButton class="mt-3 submit-button" :isDisabled="isDisabledButton" :content="'Hủy đăng ký'"
+                  :color="'rgb(255 57 57)'" @click.prevent="deletee" />
               </div>
             </form>
           </div>
@@ -107,6 +120,7 @@
 </template>
   
 <script setup>
+import {BIconChevronRight} from 'bootstrap-icons-vue';
 
 definePageMeta({
   layout: 'page',
@@ -115,11 +129,10 @@ definePageMeta({
 const route = useRoute();
 const isDisabledButton = ref(false);
 const showConfirmError = ref(false);
-const { errorAlert } = useAlert();
+const { errorAlert, successAlert } = useAlert();
 const statusShow = ref(0);
-const groupId = ref({
-  group_id: '',
-})
+const beforeShow = ref(0);
+
 const group = ref({
   id: '',
   faculty: '',
@@ -165,10 +178,10 @@ const {
   requireAuth: true,
   disableHandleErrorUnauthorized: false,
 })(
-  `/member/${route.params.id}/register`,
+  `/groups/${route.params.id}/register-mentor`,
   { immediate: false },
 )
-// đăng ký là memtor
+// đăng ký là mentor
 const {
   data: dataMentor,
   post: postMentor,
@@ -181,20 +194,32 @@ const {
   `/groups/${route.params.id}/register`,
   { immediate: false },
 )
-
-// Lấy thông tin user
+// update mentor
 const {
-  data: dataGetMe,
-  get: getMe,
-  onFetchResponse: getMeResponse,
-  onFetchError: getMeError,
+  data: dataMentorPut,
+  put: putMentor,
+  onFetchResponse: putMentorRes,
+  onFetchError: putMentorErr,
 } = useFetchApi({
   requireAuth: true,
   disableHandleErrorUnauthorized: false,
 })(
-  '/user',
-  {immediate: false},
-);
+  `/groups/${route.params.id}/register`,
+  { immediate: false },
+)
+// Xóa mentor
+const {
+  data: dataMentordel,
+  delete: delMentor,
+  onFetchResponse: delMentorRes,
+  onFetchError: delMentorErr,
+} = useFetchApi({
+  requireAuth: true,
+  disableHandleErrorUnauthorized: false,
+})(
+  `/groups/${route.params.id}/register`,
+  { immediate: false },
+)
 
 // Lấy thông tin cv của user với môn đó
 const {
@@ -206,26 +231,16 @@ const {
   requireAuth: true,
   disableHandleErrorUnauthorized: false,
 })(
-  `/member/${route.params.id}/register`,
+  `/mentor-information/${route.params.id}`,
   { immediate: false },
 )
-
-getMeResponse(() => {
-  if(group.value.mentors.find(mentor => mentor.id === dataGetMe.value.data.id)) {
-    statusShow.value = 2;
-  } else {
-    statusShow.value = 1;
-  }
-});
-getMeError(() => {
-  // deleteToken();
-});
 
 getGroup().json().execute();
 getGroupRes(() => {
   group.value = dataGetGroup.value.data.data
-  if(group.value.status === 2) {
-    getMe().json().execute();
+  // kiểm tra thực sự nhóm đang tìm mentor k hay nhập bừa id
+  if (group.value.status === 2) {
+    getCv().json().execute();
   }
   else {
     alert("Nhóm hiện không tìm người hướng dẫn!");
@@ -233,18 +248,56 @@ getGroupRes(() => {
   }
 });
 
-getCv().json().execute();
+// Đã đăng ký mentor
 getCvRes(() => {
-  // register_inform.value.cv_link = dataCv.value.data.cv_link;
-  // register_inform.value.smart_banking = dataCv.value.data.smart_banking;
+  statusShow.value = 2;
+  register_inform.value = dataCv.value.data;
+});
+// Chưa có đăng ký mentor cho nhóm này
+getCvErr(() => {
+  statusShow.value = 1;
+  // Lấy cv đăng ký mentor cho môn này chưa
+  getCvSubject().json().execute();
 });
 
 postMentorRes(() => {
+  successAlert('Bạn đã đăng ký thành công!')
+  statusShow.value = 2;
   getGroup().json().execute();
 });
 postMentorErr(() => {
   errorAlert(dataMentor.value.meta.error_message);
 })
+
+putMentorRes(() => {
+  successAlert('Chỉnh sửa thông tin thành công!');
+  register_inform.value = dataMentorPut.value.data;
+})
+putMentorErr(() => {
+  errorAlert(dataMentorPut.value.meta.error_message);
+})
+
+delMentorRes(() => {
+  successAlert('Hủy đăng ký thành công!');
+  statusShow.value = 1;
+  register_inform.value.cv_link = '';
+  register_inform.value.schedule = '';
+  register_inform.value.smart_banking = '';
+  register_inform.value.note = '';
+  register_inform.value.confirm = 'not_agreed';
+
+  getCvSubject().json().execute();
+})
+delMentorErr(() => {
+  errorAlert(dataMentordel.value.meta.error_message)
+})
+// Đã đăng ký 
+getCvSubjectRes(() => {
+  beforeShow.value = 1;
+  register_inform.value.cv_link = dataCvSubject.value.data.cv_link;
+  register_inform.value.smart_banking = dataCvSubject.value.data.smart_banking;
+});
+
 const submit = () => {
   isDisabledButton.value = true;
   showConfirmError.value = false;
@@ -255,6 +308,23 @@ const submit = () => {
     postMentor(register_inform.value).json().execute();
     isDisabledButton.value = false;
   }
+}
+const update = () => {
+  isDisabledButton.value = true;
+  showConfirmError.value = false;
+  if (register_inform.value.confirm !== 'agreed') {
+    showConfirmError.value = true;
+    isDisabledButton.value = false;
+  } else {
+    putMentor(register_inform.value).json().execute();
+    isDisabledButton.value = false;
+  }
+}
+const deletee = () => {
+  isDisabledButton.value = true;
+  showConfirmError.value = false;
+  delMentor().json().execute();
+  isDisabledButton.value = false;
 }
 </script>
   
@@ -273,14 +343,20 @@ img {
 }
 
 h4 {
-  font-size: 25px;
+  font-size: 30px;
   font-weight: 700;
   color: rgb(0, 13, 154);
 }
 
 h5 {
   font-size: 22px;
-  color: rgb(0, 85, 119);
+  color: rgb(6, 101, 138);
+  font-weight: 600;
+}
+h5:last-child {
+  font-size: 20px;
+  color: rgb(6, 101, 138);
+  font-weight: 100;
 }
 
 .group-infor span {
@@ -292,6 +368,10 @@ h5 {
 .register {
   background-color: #dfe6ec;
   padding: 20px;
+}
+
+.submit-button {
+  display: inline-block;
 }
 
 .submit-button>>>button {
@@ -313,12 +393,26 @@ label span {
   color: red;
   font-size: 13px;
 }
+
 .notice-success span {
   font-weight: 600;
   color: green;
 }
+
+.notice-success span:last-child {
+  color: rgb(0, 0, 0);
+  display: block;
+  font-size: large;
+}
+
 .quantity span {
   font-weight: 600;
+}
+.beforeShow {
+  display: block;
+  font-size: 13px;
+  padding-bottom: 10px;
+  color: rgb(0, 0, 0);
 }
 </style>
